+++ src/main/java/es/uma/khaos/docking_service/model/dlg/AutoDockSolution.java
@@ -18,7 +18,7 @@ public class AutoDockSolution {
 	/**
 	 * Estimated Free Energy of Binding
 	 */
-	private Double totalEnergy;
+	private Float totalEnergy;
 	
 	/**
 	 * Estimated Inhibition Constant (M (molar))
@@ -28,12 +28,12 @@ public class AutoDockSolution {
 	/**
 	 * (1) Final Intermolecular Energy
 	 */
-	private Double energy1;
+	private Float energy1;
 	
 	/**
 	 * (2) Final Total Internal Energy
 	 */
-	private Double energy2;
+	private Float energy2;
 	
 	/**
 	 * RMSD
@@ -42,7 +42,7 @@ public class AutoDockSolution {
 	
 	private Conformation conformation;
 	
-	public AutoDockSolution(Double totalEnergy) {
+	public AutoDockSolution(Float totalEnergy) {
 		this.totalEnergy = totalEnergy;
 		this.ki = null;
 		this.energy1 = null;
@@ -51,7 +51,7 @@ public class AutoDockSolution {
 		this.conformation = null;
 	}
 	
-	public AutoDockSolution(Double totalEnergy, Scientific ki, Double energy1, Double energy2) {
+	public AutoDockSolution(Float totalEnergy, Scientific ki, Float energy1, Float energy2) {
 		this.totalEnergy = totalEnergy;
 		this.ki = ki;
 		this.energy1 = energy1;
@@ -60,7 +60,7 @@ public class AutoDockSolution {
 		this.conformation = null;
 	}
 	
-	public AutoDockSolution(Double totalEnergy, Scientific ki, Double energy1, Double energy2, Double rmsd) {
+	public AutoDockSolution(Float totalEnergy, Scientific ki, Float energy1, Float energy2, Double rmsd) {
 		this.totalEnergy = totalEnergy;
 		this.ki = ki;
 		this.energy1 = energy1;
@@ -69,11 +69,11 @@ public class AutoDockSolution {
 		this.conformation = null;
 	}
 
-	public Double getTotalEnergy() {
+	public Float getTotalEnergy() {
 		return totalEnergy;
 	}
 
-	public void setTotalEnergy(Double totalEnergy) {
+	public void setTotalEnergy(Float totalEnergy) {
 		this.totalEnergy = totalEnergy;
 	}
 
@@ -85,19 +85,19 @@ public class AutoDockSolution {
 		this.ki = ki;
 	}
 
-	public Double getEnergy1() {
+	public Float getEnergy1() {
 		return energy1;
 	}
 
-	public void setEnergy1(Double energy1) {
+	public void setEnergy1(Float energy1) {
 		this.energy1 = energy1;
 	}
 
-	public Double getEnergy2() {
+	public Float getEnergy2() {
 		return energy2;
 	}
 
-	public void setEnergy2(Double energy2) {
+	public void setEnergy2(Float energy2) {
 		this.energy2 = energy2;
 	}
